use groups in generators
