+++ src/map-styler.ts
@@ -5,6 +5,7 @@ export {
   Generator,
   GeneratorConfig,
 } from './layer-composer/types'
+
 export { default, DEFAULT_CONFIG } from './layer-composer'
 
 export {
@@ -13,3 +14,5 @@ export {
   HEATMAP_GEOM_TYPES,
   HEATMAP_COLOR_RAMPS,
 } from './layer-composer/generators'
+
+export { default as sort, convertLegacyGroups } from './sort'
