+++ src/sort/index.ts
@@ -1,0 +1,1 @@
+export { default, convertLegacyGroups } from './sort'
