--- src/sort/types.ts
@@ -1,12 +1,0 @@
-export enum Group {
-  Background = 'background', // Solid bg color
-  Basemap = 'basemap', // Satellite tiles
-  Heatmap = 'heatmap', // Fill/gradient-based heatmaps
-  OutlinePolygons = 'outlinePolygons', // Conbtext layers with an outlined/hollow style such as EEZ, RFMOs, etc
-  Default = 'default', // Default stack position when f0roup is not specified
-  Point = 'point', // Events, etc
-  Track = 'track', // Tracks
-  BasemapForeground = 'BasemapForeground', // Graticule labels, bathymetry labels, etc
-  Label = 'label', // All non-basemap layers labels
-  Overlay = 'overlay', // Popups, ruler tool, etc
-}
